requirements.txt
Flask==3.1.0
Werkzeug==3.1.3
python-dotenv==1.0.0
Pillow==11.1.0
pytest==8.0.0
pytest-flask==1.3.0
